requirements.txt
# YouTube Shorts Agent V2 - Dependencies

# Core
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# Video Processing
moviepy>=2.0.0
Pillow>=9.0.0

# Audio
gtts>=2.3.0
pydub>=0.25.0

# Trend Research
pytrends>=4.9.0

# Web Dashboard
flask>=2.0.0

# YouTube Upload (optional)
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Performance
concurrent-log-handler>=0.9.0
